+++ src/main/java/com/andersmmg/cityessentials/datagen/ModModelProvider.java
@@ -20,5 +20,6 @@ public class ModModelProvider extends FabricModelProvider {
     @Override
     public void generateItemModels(ItemModelGenerator itemModelGenerator) {
         itemModelGenerator.register(ModItems.PAPER_BAG, Models.GENERATED);
+        itemModelGenerator.register(ModItems.GROCERY_BAG, Models.GENERATED);
     }
 }