+++ src/main/java/com/andersmmg/cityessentials/item/ModItems.java
@@ -1,6 +1,7 @@
 package com.andersmmg.cityessentials.item;
 
 import com.andersmmg.cityessentials.CityEssentials;
+import com.andersmmg.cityessentials.item.custom.GroceryBagItem;
 import com.andersmmg.cityessentials.item.custom.PaperBagItem;
 import io.wispforest.owo.itemgroup.OwoItemSettings;
 import net.minecraft.item.Item;
@@ -9,6 +10,7 @@ import net.minecraft.registry.Registry;
 
 public class ModItems {
     public static final Item PAPER_BAG = registerItem("paper_bag", new PaperBagItem(new OwoItemSettings().maxCount(1).group(ModItemGroups.CITYESSENTIALS_GROUP)));
+    public static final Item GROCERY_BAG = registerItem("grocery_bag", new GroceryBagItem(new OwoItemSettings().maxCount(1).group(ModItemGroups.CITYESSENTIALS_GROUP)));
 
     private static Item registerItem(String name, Item item) {
         return Registry.register(Registries.ITEM, CityEssentials.id(name), item);